Convert from the iCloud key-value store to CloudKit.  MKiCloudSync uses the iCloud key-value store (NSUbiquitousKeyValueStore) which is only available in apps distributed through the App Store.  Some apps cannot be or don't want to be distributed through the App Store, so an alternative is needed, which is where CloudKit comes in.
Use of CloudKit requires far more code than the key-value store, and features have been added such as a key match list, only updating things that have changed, and notification back to other code which a change is received.  Since the library has changed so greatly, the name is also changed to avoid confusion and to call attention to the specific iCloud mechanism it uses and what it syncs.

This initial commit has been demonstrated to work but warrants further evaluation and testing.  It includes copious logging statements, a bit of debug code, a bit of commented-out stuff that can be removed, and a bit of commented-out stuff that should maybe be restored.  It has not yet been verified to be of consistent style.